12 Refactor Add Task component
